Share the AIFS Single coordinates between the two template configs
The materialized and virtual configs defined the same grid, forecast axis and
seven of eight coordinates. Both now extend a common base and add only what is
unique to them: ingested_forecast_length for the materialized variant, the
pressure_level dimension and coordinate for the virtual one.

Both templates regenerate byte for byte unchanged. Two statistics the variants
report differently (valid_time max, expected_forecast_length min) are fields on
the base so neither dataset's published metadata moves.
