+++ src/reformatters/ecmwf/aifs_single/template_config.py
@@ -1,0 +1,223 @@
+from collections.abc import Sequence
+from typing import Any
+
+import numpy as np
+import pandas as pd
+import xarray as xr
+from pydantic import computed_field
+
+from reformatters.common.config_models import (
+    Coordinate,
+    CoordinateAttrs,
+    Encoding,
+    StatisticsApproximate,
+)
+from reformatters.common.template_config import (
+    SPATIAL_REF_COORDS,
+    TemplateConfig,
+)
+from reformatters.common.types import AppendDim, Timedelta, Timestamp
+from reformatters.common.zarr import BLOSC_8BYTE_ZSTD_LEVEL3_SHUFFLE
+from reformatters.ecmwf.ecmwf_config_models import EcmwfDataVar
+
+
+class EcmwfAifsSingleCommonTemplateConfig[DATA_VAR: EcmwfDataVar](
+    TemplateConfig[DATA_VAR]
+):
+    """The grid, forecast axis and coordinates shared by the AIFS Single variants.
+
+    Subclasses extend `dimension_coordinates`, `derive_coordinates` and `coords` with
+    the parts unique to their layout (a vertical dimension, an extra coordinate).
+    """
+
+    append_dim: AppendDim = "init_time"
+    # Earlier data (2024-02-29 to 2024-03-31) uses a reduced Gaussian grid.
+    append_dim_start: Timestamp = pd.Timestamp("2024-04-01T00:00")
+    append_dim_frequency: Timedelta = pd.Timedelta("6h")
+
+    # Statistics the variants report differently; see each subclass.
+    valid_time_statistic_max: str
+    expected_forecast_length_statistic_min: str
+
+    def dimension_coordinates(self) -> dict[str, Any]:
+        return {
+            self.append_dim: self.append_dim_coordinates(
+                self.append_dim_start + self.append_dim_frequency
+            ),
+            "lead_time": pd.timedelta_range("0h", "360h", freq="6h"),
+            "latitude": np.flip(np.arange(-90, 90.25, 0.25)),
+            "longitude": np.arange(-180, 180, 0.25),
+        }
+
+    def derive_coordinates(
+        self, ds: xr.Dataset
+    ) -> dict[str, xr.DataArray | tuple[tuple[str, ...], np.ndarray[Any, Any]]]:
+        return {
+            "valid_time": ds["init_time"] + ds["lead_time"],
+            "expected_forecast_length": (
+                (self.append_dim,),
+                np.full(
+                    ds[self.append_dim].size,
+                    self.dimension_coordinates()["lead_time"].max(),
+                    dtype="timedelta64[us]",
+                ),
+            ),
+            "spatial_ref": SPATIAL_REF_COORDS,
+        }
+
+    @computed_field
+    @property
+    def coords(self) -> Sequence[Coordinate]:
+        dim_coords = self.dimension_coordinates()
+        append_dim_coordinate_chunk_size = self.append_dim_coordinate_chunk_size()
+
+        return [
+            Coordinate(
+                name=self.append_dim,
+                encoding=Encoding(
+                    dtype="int64",
+                    fill_value=0,
+                    compressors=[BLOSC_8BYTE_ZSTD_LEVEL3_SHUFFLE],
+                    calendar="proleptic_gregorian",
+                    units="seconds since 1970-01-01 00:00:00",
+                    chunks=append_dim_coordinate_chunk_size,
+                    shards=None,
+                ),
+                attrs=CoordinateAttrs(
+                    long_name="Forecast initialization time",
+                    standard_name="forecast_reference_time",
+                    units="seconds since 1970-01-01 00:00:00",
+                    statistics_approximate=StatisticsApproximate(
+                        min=dim_coords[self.append_dim].min().isoformat(), max="Present"
+                    ),
+                ),
+            ),
+            Coordinate(
+                name="lead_time",
+                encoding=Encoding(
+                    dtype="float64",
+                    fill_value=float("nan"),
+                    compressors=[BLOSC_8BYTE_ZSTD_LEVEL3_SHUFFLE],
+                    units="seconds",
+                    chunks=len(dim_coords["lead_time"]),
+                    shards=None,
+                ),
+                attrs=CoordinateAttrs(
+                    long_name="Forecast lead time",
+                    standard_name="forecast_period",
+                    units="seconds",
+                    statistics_approximate=StatisticsApproximate(
+                        min=str(dim_coords["lead_time"].min()),
+                        max=str(dim_coords["lead_time"].max()),
+                    ),
+                ),
+            ),
+            Coordinate(
+                name="latitude",
+                encoding=Encoding(
+                    dtype="float64",
+                    fill_value=np.nan,
+                    compressors=[BLOSC_8BYTE_ZSTD_LEVEL3_SHUFFLE],
+                    chunks=len(dim_coords["latitude"]),
+                    shards=None,
+                ),
+                attrs=CoordinateAttrs(
+                    long_name="Latitude",
+                    standard_name="latitude",
+                    units="degree_north",
+                    axis="Y",
+                    statistics_approximate=StatisticsApproximate(
+                        min=float(dim_coords["latitude"].min()),
+                        max=float(dim_coords["latitude"].max()),
+                    ),
+                ),
+            ),
+            Coordinate(
+                name="longitude",
+                encoding=Encoding(
+                    dtype="float64",
+                    fill_value=np.nan,
+                    compressors=[BLOSC_8BYTE_ZSTD_LEVEL3_SHUFFLE],
+                    chunks=len(dim_coords["longitude"]),
+                    shards=None,
+                ),
+                attrs=CoordinateAttrs(
+                    long_name="Longitude",
+                    standard_name="longitude",
+                    units="degree_east",
+                    axis="X",
+                    statistics_approximate=StatisticsApproximate(
+                        min=float(dim_coords["longitude"].min()),
+                        max=float(dim_coords["longitude"].max()),
+                    ),
+                ),
+            ),
+            Coordinate(
+                name="valid_time",
+                encoding=Encoding(
+                    dtype="int64",
+                    fill_value=0,
+                    compressors=[BLOSC_8BYTE_ZSTD_LEVEL3_SHUFFLE],
+                    calendar="proleptic_gregorian",
+                    units="seconds since 1970-01-01 00:00:00",
+                    chunks=(
+                        append_dim_coordinate_chunk_size,
+                        len(dim_coords["lead_time"]),
+                    ),
+                    shards=None,
+                ),
+                attrs=CoordinateAttrs(
+                    long_name="Valid time",
+                    standard_name="time",
+                    units="seconds since 1970-01-01 00:00:00",
+                    statistics_approximate=StatisticsApproximate(
+                        min=self.append_dim_start.isoformat(),
+                        max=self.valid_time_statistic_max,
+                    ),
+                ),
+            ),
+            Coordinate(
+                name="expected_forecast_length",
+                encoding=Encoding(
+                    dtype="float64",
+                    fill_value=float("nan"),
+                    compressors=[BLOSC_8BYTE_ZSTD_LEVEL3_SHUFFLE],
+                    units="seconds",
+                    chunks=append_dim_coordinate_chunk_size,
+                    shards=None,
+                ),
+                attrs=CoordinateAttrs(
+                    long_name="Expected forecast length",
+                    units="seconds",
+                    statistics_approximate=StatisticsApproximate(
+                        min=self.expected_forecast_length_statistic_min,
+                        max=str(dim_coords["lead_time"].max()),
+                    ),
+                ),
+            ),
+            Coordinate(
+                name="spatial_ref",
+                encoding=Encoding(
+                    dtype="int64",
+                    fill_value=0,
+                    chunks=(),
+                    shards=None,
+                ),
+                attrs=CoordinateAttrs(
+                    units=None,
+                    statistics_approximate=None,
+                    crs_wkt='GEOGCS["unknown",DATUM["unknown",SPHEROID["unknown",6371229,0]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AXIS["Longitude",EAST],AXIS["Latitude",NORTH]]',
+                    semi_major_axis=6371229.0,
+                    semi_minor_axis=6371229.0,
+                    inverse_flattening=0.0,
+                    reference_ellipsoid_name="unknown",
+                    longitude_of_prime_meridian=0.0,
+                    prime_meridian_name="Greenwich",
+                    geographic_crs_name="unknown",
+                    horizontal_datum_name="unknown",
+                    grid_mapping_name="latitude_longitude",
+                    spatial_ref='GEOGCS["unknown",DATUM["unknown",SPHEROID["unknown",6371229,0]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AXIS["Longitude",EAST],AXIS["Latitude",NORTH]]',
+                    comment="This coordinate reference system matches the source data which follows WMO conventions of assuming the earth is a perfect sphere with a radius of 6,371,229m. It is similar to EPSG:4326, but EPSG:4326 uses a more accurate representation of the earth's shape.",
+                ),
+            ),
+        ]
